Hildongui: Show more logs (button+dialog with tabs)
Basic view shows 10 logs for cache. And if there are more than 10
logs, there are button 'show more logs', which starts new dialog with
tabs containing all logs from cache, 30 logs on each.

These extra logs are not loaded to memory until button is pressed,
because it could take plenty of time when creating this dialog (if
there are hundreds of logs).